+++ phpstan-bootstrap.php
@@ -14,6 +14,10 @@ define( 'WP4ODOO_PLUGIN_BASENAME', 'wp4odoo/wp4odoo.php' );
 define( 'WP4ODOO_MIN_ODOO_VERSION', 14 );
 define( 'WPINC', 'wp-includes' );
 
+if ( ! defined( 'HOUR_IN_SECONDS' ) ) {
+	define( 'HOUR_IN_SECONDS', 3600 );
+}
+
 /**
  * Stub for the main plugin singleton.
  *